Add button to collapse/expand file body
When hound search returns a large number of results it can be annoying to
scroll through the everything to find what you need. This adds a button to
collapse file body in search results.

Note that this change depends on #324 which adds new version of React which
has support for React.createRef.